Extract schema and resolvers into separate files
