+++ server/src/resolvers.ts
@@ -1,0 +1,7 @@
+import Query from './resolvers/Query';
+
+const resolvers = {
+    Query,
+  }
+
+  export default resolvers;
